fix average to data type double

+++ app/src/main/java/scottsalvador/pons/com/ponsscottsalvadorexam2/Student.java
@@ -1,21 +1,20 @@
 package scottsalvador.pons.com.ponsscottsalvadorexam2;
 
-import android.widget.Toast;
 
 public class Student {
     String fname, lname;
-    Long exam1, exam2, avg;
+    double exam1, exam2, avg;
 
     public Student(){
 
     }
 
-    public Student(String fname, String lname, Long exam1, Long exam2) {
+    public Student(String fname, String lname, double exam1, double exam2) {
         this.fname = fname;
         this.lname = lname;
         this.exam1 = exam1;
         this.exam2 = exam2;
-        this.avg = (long)((exam1+exam2)/2);
+        this.avg = (exam1+exam2)/2;
     }
 
     public String getFname() {
@@ -26,15 +25,15 @@ public class Student {
         return lname;
     }
 
-    public Long getExam1() {
+    public Double getExam1() {
         return exam1;
     }
 
-    public Long getExam2() {
+    public Double getExam2() {
         return exam2;
     }
 
-    public Long getAVG() {
+    public Double getAVG() {
         return avg;
     }
 }
